chore: onboard layout without SC
