+++ apps/extension/src/ui/apps/onboard/routes/AddAccount/AccountAddWrapper.tsx
@@ -1,6 +1,6 @@
 import { OnboardDialog } from "@ui/apps/onboard/components/OnboardDialog"
 import { useOnboard } from "@ui/apps/onboard/context"
-import { Layout } from "@ui/apps/onboard/layout"
+import { OnboardLayout } from "@ui/apps/onboard/OnboardLayout"
 
 export const AccountAddWrapper = ({
   title,
@@ -14,11 +14,11 @@ export const AccountAddWrapper = ({
   const { setOnboarded } = useOnboard()
 
   return (
-    <Layout withBack>
+    <OnboardLayout withBack>
       <OnboardDialog className="w-[68rem]" title={title}>
         {subtitle && <p className="text-body-secondary mb-8 text-base">{subtitle}</p>}
         {render(setOnboarded)}
       </OnboardDialog>
-    </Layout>
+    </OnboardLayout>
   )
 }
